requirements.txt
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.15.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.0
oauthlib==3.1.0
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.13.0
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
